updated templates and added static files

+++ dynamite_nsm/agent_api/blueprints/admin/users.py
@@ -1,7 +1,7 @@
 from flask_security import roles_accepted
 from sqlalchemy.exc import IntegrityError
 from flask import render_template, Blueprint
-from flask import request, redirect, url_for, flash
+from flask import request, redirect, url_for
 from flask_login import logout_user, current_user
 
 from flask_security import SQLAlchemySessionUserDatastore
@@ -74,8 +74,9 @@ def create_new_user_form():
             return redirect(url_for('users.create_user_form_html'))
     except KeyError:
         redirect(url_for('users.create_user_form_html'))
-    user_obj = user_datastore.find_user(email=current_user.email)
-    user_datastore.delete_user(user_obj)
-    db_session.commit()
-    logout_user()
+    if current_user.email == 'admin@dynamite.local':
+        user_obj = user_datastore.find_user(email=current_user.email)
+        user_datastore.delete_user(user_obj)
+        db_session.commit()
+        logout_user()
     return redirect('/home')
